backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
openai>=1.3.0
python-dotenv>=1.0.0
httpx>=0.25.2
pymongo>=4.6.0
motor>=3.3.2
numpy>=1.21.0
requests>=2.31.0
python-multipart>=0.0.20
PyPDF2>=3.0.1
python-magic>=0.4.27
sentence-transformers>=2.2.2
torch>=2.0.0
pdfminer.six>=20240706
rank-bm25>=0.2.2
langchain-text-splitters>=0.2.0
langchain>=0.2.16
langchain-community>=0.2.16
langchain-mongodb>=0.2.0
